requirements.txt
# General
stringcase # used eg to format strings snake_case
lxml
xmltodict==0.11.0
pycrypto==2.6.1

# Dev tools
isort

# Scraping tools
beautifulsoup4
wget
scrapy
zeep==3.2.0
pandaSDMX==0.9


# File storage
boto3==1.9.5

# Template and export
jinja2==2.10
plotly==3.4.0
psutil==5.4.8 # required for plotly image export

# Django
django==2.1.3
django-material==1.4.1
django-suit==0.2.26
django-crispy-forms==1.7.2
django-environ==0.4.4
django-redis==4.5.0
django-tinymce4-lite==1.7.2
django-storages==1.6.6 # Access AWS S3
django-bootstrap-datepicker-plus==3.0.4
#django_celery_beat==1.2.0
git+https://github.com/celery/django-celery-beat # forced to do this due to a race condition bug in celery beat 1.2.0
django-celerybeat-status==0.0.10
django-simple-history==2.5.1
django-debug-toolbar==1.10.1

# Document management
git+https://github.com/pdftables/python-pdftables-api.git
diff_match_patch # diff compare two documents
pyenchant==2.0.0
python-magic==0.4.15
libmagic==1.0
reportlab==3.5.9
xhtml2pdf==0.2.3
pdfrw==0.4

# Data
pandas==0.23.4

# IO and background tasks
redis==2.10.6
gunicorn==19.8.1
celery==4.2.1
flower==0.9.2
pysftp==0.2.9
exchangelib==1.12.0
mailchimp3==3.0.6

# Google
google-api-python-client==1.7.6
google-auth==1.6.1
google-auth-httplib2==0.0.3

# Database
psycopg2-binary==2.7.4 # Postgresql
SQLAlchemy==1.2.4
sqlalchemy-migrate==0.11.0
sqlalchemy-views==0.2.1
sqlalchemy-utils==0.33.1

# Dates
pytz==2018.3 # World timezone definitions
python-dateutil==2.7.3
arrow==0.12.1 # date management

# Documentation
sphinx==1.8.0
sphinx_rtd_theme==0.4.1

# Pygleif
git+git://github.com/ggravlingen/pygleif.git

# NCR specific | private
# Use keys instead
git+ssh://git@github.com/NordicCreditRating/pycreditrating.git
git+ssh://git@github.com/NordicCreditRating/pyesmaradar.git
git+ssh://git@github.com/NordicCreditRating/pyfindata.git
